requirements.txt
# Core
pandas>=2.0
polars>=0.20
numpy>=1.24

# ML
scikit-learn>=1.3
umap-learn>=0.5
hdbscan>=0.8
shap>=0.42

# LLM — Azure AI Foundry + Azure OpenAI
openai>=1.0
azure-core

# Vector Store
faiss-cpu>=1.7

# Dashboard
streamlit>=1.28
plotly>=5.15
altair>=5.0

# Utils
python-dotenv>=1.0
pyarrow>=14.0
joblib>=1.3

# Pipeline (opcional futuro)
prefect>=2.0
